tcl: modernize + add tcl/8.6.11
