--- recipes/tcl/8.6.10/conandata.yml
@@ -1,8 +1,0 @@
-sources:
-  8.6.10:
-    sha256: 5196dbf6638e3df8d5c87b5815c8c2b758496eb6f0e41446596c9a4e638d87ed
-    url: https://downloads.sourceforge.net/project/tcl/Tcl/8.6.10/tcl8.6.10-src.tar.gz
-patches:
-  8.6.10:
-    - patch_file: patches/no-read-only-data.patch
-      base_path: source_subfolder
