+++ recipes/tcl/all/conandata.yml
@@ -1,0 +1,14 @@
+sources:
+  8.6.11:
+    url: "https://downloads.sourceforge.net/project/tcl/Tcl/8.6.11/tcl8.6.11-src.tar.gz"
+    sha256: "8c0486668586672c5693d7d95817cb05a18c5ecca2f40e2836b9578064088258"
+  8.6.10:
+    url: "https://downloads.sourceforge.net/project/tcl/Tcl/8.6.10/tcl8.6.10-src.tar.gz"
+    sha256: "5196dbf6638e3df8d5c87b5815c8c2b758496eb6f0e41446596c9a4e638d87ed"
+patches:
+  8.6.11:
+    - patch_file: "patches/0001-8.6.11-no-read-only-data.patch"
+      base_path: "source_subfolder"
+  8.6.10:
+    - patch_file: "patches/0001-8.6.10-no-read-only-data.patch"
+      base_path: "source_subfolder"
