+++ recipes/tcl/all/patches/0001-8.6.11-no-read-only-data.patch
@@ -1,11 +1,9 @@
-diff --git a/tools/installData.tcl b/tools/installData.tcl
-index 4b43f1e..3048dba 100644
---- a/tools/installData.tcl
-+++ b/tools/installData.tcl
-@@ -33,18 +33,13 @@ proc copyDir {d1 d2} {
+--- tools/installData.tcl
++++ tools/installData.tcl
+@@ -33,18 +33,13 @@
  	    file copy -force $f [file join $d2 $ftail]
  	    if {$::tcl_platform(platform) eq {unix}} {
- 		file attributes [file join $d2 $ftail] -permissions 0644
+ 		file attributes [file join $d2 $ftail] -permissions 0o644
 -	    } else {
 -		file attributes [file join $d2 $ftail] -readonly 1
  	    }
@@ -13,7 +11,7 @@ index 4b43f1e..3048dba 100644
      }
  
      if {$::tcl_platform(platform) eq {unix}} {
- 	file attributes $d2 -permissions 0755
+ 	file attributes $d2 -permissions 0o755
 -    } else {
 -	file attributes $d2 -readonly 1
      }
