+++ recipes/tcl/config.yml
@@ -1,0 +1,5 @@
+versions:
+  "8.6.11":
+    folder: "all"
+  "8.6.10":
+    folder: "all"
